Reorganise components into `Layout` folder
